Containerisation of application and updating Database Connection (#29)
## What I've Done
I have cleaned up the Dockerfile, as well as update the Database
connection to look for Environmental Variables if they are there. If
they aren't it defaults to the appsettings.json. This is for Cloud Run
so we can use secrets as environmental variables and connect to the
database, as we don't want the details to be visible. I've also deleted
files that I don't think we will need.